% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestimate.R
\name{bestimate}
\alias{bestimate}
\title{Bayesian finite population with BART and FPBB}
\usage{
bestimate(samp, ref, y_var_names, x_var_names, sp_wts, propensity = TRUE,
  prediction = TRUE, double_robust = TRUE, dr_propensity_transform = odds,
  decompose_bias = TRUE, pred_subsample_size = 10000,
  posterior_draws = 1000, model_params = list(mc.cores = 1, ntree = 50),
  seed = 12345)
}
\arguments{
\item{samp}{\code{data.frame} containing the survey sample data.}

\item{ref}{\code{data.frame} containing the reference sample.}

\item{y_var_names}{Character vector containing names of the outcome
variables.}

\item{x_var_names}{Character vector containing names of the covariates}

\item{posterior_draws}{Number of posterior draws}

\item{sp_wt_frame}{\code{data.frame} containing one column of weights
for each synthetic population.}
}
\value{
Good question
}
\description{
Bayesian finite population with BART and FPBB
}
