% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart_functions.R
\name{dummify}
\alias{dummify}
\title{Convert factors in data frame to binary variables}
\usage{
dummify(df, drop_unused = TRUE, sep = "_")
}
\arguments{
\item{df}{Data frame to convert}

\item{drop_unused}{Whether or not to drop levels that do not have any
corresponding observations.}

\item{sep}{Separator between variable and value names}
}
\value{
Returns a tibble where the factors have been converted to binary
variables
}
\description{
Convert factors in data frame to binary variables
}
