% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_propensities.R
\name{estimate_propensities}
\alias{estimate_propensities}
\title{Title}
\usage{
estimate_propensities(sp_wt, sp_id, x_samp, x_ref, bart_params,
  num_replicates = 10)
}
\arguments{
\item{sp_id}{}

\item{x_samp}{}

\item{x_ref}{}

\item{bart_params}{}

\item{subsamp_ids}{}
}
\description{
Title
}
