% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestimate.R
\name{expand_sp_weights}
\alias{expand_sp_weights}
\title{Expand synthetic population weights}
\usage{
expand_sp_weights(wts)
}
\arguments{
\item{wts}{Set of integer weights to expand}
}
\value{
Returns a vector where the index of each unit in the reference
sample is repeated \code{w} times where \code{w} is its weight.
}
\description{
Expand synthetic population weights
}
