% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpbb.R
\name{fpbb_synth_pops}
\alias{fpbb_synth_pops}
\title{Create synthetic populations with FPBB.}
\usage{
fpbb_synth_pops(weights, L = 2, N = length(weights) * 2,
  return_weights = TRUE)
}
\arguments{
\item{weights}{Vector of weights for each unit in the reference sample. 
These determine the initial probabilities of selection for the weighted
Polya urn scheme used to generate each synthetic population.}

\item{L}{Number of synthetic populations to create.}

\item{N}{Size of the synthetic populations. In theory this is the size of the 
full population but in practice only needs to be 10-20 times as large as
the reference sample.}

\item{return_weights}{Logical flag indicating whether the synthetic
populations should be returned as weights or indices. 
\enumerate{
  \item If \code{TRUE}, output will contain the number of times each unit 
  in the reference sample appears in the synthetic population. 
  \item If \code{FALSE} output will be a vector containing the indices of 
  each unit in the order they were selected. This permits the use of 
  synthetic populations with procedure that cannot handle weights.
  }}
}
\value{
Returns a data frame with one column for each synthetic population.
  If \code{return_weights = TRUE} each column will be a vector containing
  the number of times a unit appears in the synthetic population. If 
  \code{return_weights = FALSE} each column will be a vector containing
  the row number of each unit in the reference sample in the order it was
  selected.
}
\description{
Create synthetic populations with FPBB.
}
