% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestimate.R
\name{odds}
\alias{odds}
\title{Calculate odds of a proportion}
\usage{
odds(p)
}
\arguments{
\item{p}{A proportion}
}
\value{
Returns the odds of p as \code{p/(1-p)}
}
\description{
Calculate odds of a proportion
}
