% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart_functions.R
\name{pbart_posterior}
\alias{pbart_posterior}
\title{Get posterior samples from pbart object for newdata}
\usage{
pbart_posterior(pbart_fit, newdata, mc.cores = 1,
  return_posterior_mean = FALSE)
}
\arguments{
\item{pbart_fit}{Object of class "\code{pbart}"}

\item{newdata}{Data frame containing X variables for use in generating
new predictions}

\item{mc.cores}{Number of cores to use}

\item{return_posterior_mean}{If TRUE returns the posterior mean for each
observation in \code{newdata}. Otherwise returns the full set of posterior
draws.  Defaults to \code{FALSE}.}
}
\value{
Returns a tibble containing predicted values. Either a single
posterior mean or the full set of posterior samples.
}
\description{
Get posterior samples from pbart object for newdata
}
