% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestimate.R
\name{posterior_means}
\alias{posterior_means}
\title{Calculate the posterior distribution for Y over a matrix of weights}
\usage{
posterior_means(y, w)
}
\arguments{
\item{y}{A data frame or matrix containing one or more columns of Y variables}

\item{w}{A data frame or matrix containing one or more columns of weights}
}
\value{
Returns a weighted mean for each combination of \code{y} by \code{x}.
}
\description{
Calculate the posterior distribution for Y over a matrix of weights
}
