% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{prop2weight}
\alias{prop2weight}
\title{Convert propensity to weight}
\usage{
prop2weight(p, p_denom = TRUE)
}
\arguments{
\item{p}{A vector of propensity scores}

\item{p_denom}{If \code{TRUE}, indicates that the weights should
be calculated as \code{wt = (1 - p)/p}. Otherwise \code{wt = p/(1 - p)}}
}
\value{
A vector of propensity weights
}
\description{
Takes a vector of propensity scores for a sample and
converts them into propensity weights using their
odds. By default, it assumes that the propensities
represent the probability than an observation comes from
the sample rather than the reference dataset. If the propensity
is the probability of being from the reference dataset, set
\code{p_denom = FALSE}.
}
