% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bart_fit.R
\name{run_bart_fit}
\alias{run_bart_fit}
\title{Wrapper to run a bart model and return only posterior distributions}
\usage{
run_bart_fit(y_train, x_train, x_test, bart_params, seed = NULL)
}
\arguments{
\item{y_train}{}

\item{x_train}{}

\item{x_test}{}

\item{bart_params}{}

\item{seed}{}
}
\description{
Wrapper to run a bart model and return only posterior distributions
}
