% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{sp_subsample}
\alias{sp_subsample}
\title{Create subsample from a data.frame containing synthetic populations}
\usage{
sp_subsample(synth_pops, size, as_weights = TRUE, seed = NULL)
}
\arguments{
\item{synth_pops}{\code{data.frame} containing synthetic populations}

\item{size}{Size of the subsamples}

\item{as_weights}{Indicates that the synthetic populations are 
represented as weights. If FALSE, synthetic populations are 
indices of records in the reference dataset.}

\item{seed}{Random seed}
}
\value{
A list containing subsamples for each synthetic population
}
\description{
Create subsample from a data.frame containing synthetic populations
}
