% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestimate.R
\name{bestimate}
\alias{bestimate}
\title{Bayesian finite population with BART and FPBB}
\usage{
bestimate(samp, ref, y_var_names, x_var_names, sp_wts = NULL,
  propensity = TRUE, prediction = TRUE, double_robust_wt = TRUE,
  double_robust_reg = TRUE, dr_propensity_transform = log,
  propensity_replicates = 1, quality_measures = TRUE,
  pred_subsample_size = 10000, posterior_draws = 1000,
  bart_params = list(mc.cores = 1, ntree = 50))
}
\arguments{
\item{samp}{\code{data.frame} containing the survey sample data.}

\item{ref}{\code{data.frame} containing the reference sample.}

\item{y_var_names}{Character vector containing names of the outcome
variables.}

\item{x_var_names}{Character vector containing names of the covariates}

\item{posterior_draws}{Number of posterior draws}

\item{sp_wt_frame}{\code{data.frame} containing one column of weights
for each synthetic population.}
}
\value{
Good question
}
\description{
Bayesian finite population with BART and FPBB
}
