% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_propensities.R
\name{estimate_propensities}
\alias{estimate_propensities}
\title{Title}
\usage{
estimate_propensities(subsamp_ids, sp_id, x_samp, x_ref, bart_params,
  seed = NULL)
}
\arguments{
\item{subsamp_ids}{}

\item{sp_id}{}

\item{x_samp}{}

\item{x_ref}{}

\item{bart_params}{}

\item{seed}{}
}
\description{
Title
}
